Target .NET 5.0 (#166)
v1 targeted .NET Standard 2.0.  I was originally planning having v2 target .NET Standard 2.1, but I think going right to .NET 5 is the move.  .NET 5 is set to be released in November of this year.

* .NET 5 adds `T?`, which lets you do nullable references in a lot more places with generics.  Of course, there's a lot of generics here.
* [This blog post](https://devblogs.microsoft.com/dotnet/the-future-of-net-standard/) on the future of .NET Standard recommends going right to .NET 5:

> Most code can probably skip .NET Standard 2.1 and go straight to .NET 5

Upgrading to .NET 5 also exposed a bunch of nullability errors I wasn't getting before.  It probably has something to do with going from the attributes to `T?`.

Other changes:
* Fix potential NREs in `MappedComparer`, `MappedEqualityComparer`, and `Token`